+++ gitsmartcommit/__init__.py
@@ -1,3 +1,3 @@
 from .cli import main as cli_main
 
-__version__ = "0.1.1"
+__version__ = "0.1.2"